Photo Uploading on Order Completion
When an order is updated, if it has not yet been submitted, and it has
prints and a successful payment, the photos for the order will be
transferred to Madera.

Although photos are being transfered, we still need to accept payments,
and also to create some sort of manifest that Madera will use to
determine quantities and shipping address.